+++ ops/ops.py
@@ -11,14 +11,38 @@ from bpy_extras.io_utils import (
         ExportHelper,
         )
 
+from . import export_trk
+
+if "bpy" in locals():
+    import importlib
+    importlib.reload(export_trk)
+
 class EXPORT_OT_cdtrk(bpy.types.Operator, ExportHelper):
     bl_idname       = 'export_scene.cdtrk'
     bl_label        = 'Export TRK'
     bl_description  = 'Export Crashday RE .trk map'
+    bl_options      = {'UNDO'}
 
     filename_ext    = '.trk'
-    filter_globals  : StringProperty(default='*.trk',
-                                        options={'HIDDEN'})
+    filter_glob     : StringProperty(default='*.trk',
+                                     options={'HIDDEN'})
+
+    use_selection   : BoolProperty(
+        name        = "Selection Only",
+        description = "Export selected objects only",
+        default     = False
+    )
+
+    use_mesh_modifiers: BoolProperty(
+        name        = "Apply Mesh Modifiers",
+        description = "Apply modifiers",
+        default     = True
+    )
 
     def execute(self, context):
+        keywords = self.as_keywords(ignore=('filter_glob',
+                                            "check_existing",))
+
+        export_trk.export_trk(self, context, **keywords)
+
         return {'FINISHED'}