+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+TODO:
+- add limits to the trk options
+- APPLY ROTATION N SCALE
+- deal with selection only stuff
+
 ## [Unreleased]
 ### Added
 ### Changed